requirements.txt
numpy
scipy
numpy-stl
